Unburden topic_tools.__init__.py
