+++ kalite/coachreports/models.py
@@ -7,7 +7,7 @@ from django.core.exceptions import ObjectDoesNotExist
 from kalite.facility.models import FacilityUser
 from kalite.main.models import ExerciseLog, VideoLog
 from kalite.playlist.models import VanillaPlaylist as Playlist, QuizLog
-from kalite.topic_tools import get_slug2id_map, get_id2slug_map, convert_leaf_url_to_id, get_leafed_topics, get_content_cache, get_exercise_cache
+from kalite.topic_tools.api import get_slug2id_map, get_id2slug_map, convert_leaf_url_to_id, get_leafed_topics, get_content_cache, get_exercise_cache
 
 class PlaylistProgressParent:
     """Parent class for helpful class methods"""
