+++ kalite/control_panel/views.py
@@ -31,7 +31,7 @@ from kalite.facility.forms import FacilityForm
 from kalite.facility.models import Facility, FacilityUser, FacilityGroup
 from kalite.main.models import ExerciseLog, VideoLog, UserLog, UserLogSummary
 from kalite.shared.decorators.auth import require_authorized_admin, require_authorized_access_to_student_data
-from kalite.topic_tools import get_exercise_cache
+from kalite.topic_tools.api import get_exercise_cache
 from kalite.version import VERSION, VERSION_INFO
 
 # TODO(dylanjbarth): this looks awful
