+++ kalite/distributed/tests/browser_tests/distributed.py
@@ -21,7 +21,7 @@ from kalite.testing.base import KALiteBrowserTestCase
 from kalite.testing.mixins.browser_mixins import BrowserActionMixins
 from kalite.testing.mixins.django_mixins import CreateAdminMixin
 from kalite.testing.mixins.facility_mixins import FacilityMixins, CreateFacilityMixin
-from kalite.topic_tools import get_node_cache
+from kalite.topic_tools.api import get_node_cache
 
 logging = settings.LOG
