+++ kalite/main/api_resources.py
@@ -8,7 +8,7 @@ from django.conf import settings
 from .models import VideoLog, ExerciseLog, AttemptLog, ContentLog
 
 from kalite.distributed.api_views import get_messages_for_api_calls
-from kalite.topic_tools import get_exercise_data, get_assessment_item_data, get_content_data
+from kalite.topic_tools.api import get_exercise_data, get_assessment_item_data, get_content_data
 from kalite.shared.api_auth.auth import UserObjectsOnlyAuthorization
 from kalite.facility.api_resources import FacilityUserResource
