+++ kalite/main/api_views.py
@@ -9,7 +9,7 @@ from fle_utils.internet.decorators import api_handle_error_with_json
 from fle_utils.internet.classes import JsonResponse
 from fle_utils.internet.webcache import backend_cache_page
 
-from kalite.topic_tools import get_flat_topic_tree, get_topic_tree
+from kalite.topic_tools.api import get_flat_topic_tree, get_topic_tree
 
 @api_handle_error_with_json
 @backend_cache_page
